improve dev env
