+++ server/package.json
@@ -5,15 +5,14 @@
   "license": "MIT",
   "scripts": {
     "dev": "tsx index.ts",
-    "check": "tsc"
+    "check": "tsc",
+    "postinstall": "npx prisma generate"
   },
   "dependencies": {
     "@prisma/client": "6.4.1",
     "express": "^4.21.2",
     "express-session": "^1.18.1",
-    "prisma": "^6.4.1",
-    "zod": "^3.23.8",
-    "zod-validation-error": "^3.4.0"
+    "prisma": "^6.4.1"
   },
   "devDependencies": {
     "@types/express": "4.17.21",
